% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sce.R
\name{altexp_load}
\alias{altexp_load}
\title{Load individual altExp}
\usage{
altexp_load(
  altExpName,
  path,
  rownames_add_prefix = TRUE,
  rownames_prefix_sep = "_"
)
}
\arguments{
\item{altExpName}{Name of altExp to be loaded}

\item{path}{Path to multisce folder for the object}

\item{rownames_add_prefix}{Should rowname prefix (altExp name) be added?}

\item{rownames_prefix_sep}{Separator used for pasting rownames prefix}
}
\description{
Load individual altExp
}
