% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sce.R
\name{altexp_save}
\alias{altexp_save}
\title{Save individual altExp}
\usage{
altexp_save(
  altExpName,
  sce,
  path,
  rownames_strip_prefix = TRUE,
  rownames_prefix_sep = "_"
)
}
\arguments{
\item{altExpName}{Name of altExp to be saved}

\item{sce}{SingleCellExperiment object}

\item{path}{Path to multisce folder for the object}

\item{rownames_strip_prefix}{Should rowname prefix (altExp name) be stripped before saving?}

\item{rownames_prefix_sep}{Separator used for pasting rownames prefix}
}
\description{
Save individual altExp
}
