% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/barcodes.R
\name{barcodes_check}
\alias{barcodes_check}
\title{Check barcode data}
\usage{
barcodes_check(sce, path = multisce_path(sce), bc_filename = "barcodes.tsv")
}
\arguments{
\item{sce}{SingleCellExperiment object}

\item{path}{Path to multisce folder for the object}

\item{bc_filename}{Name of barcodes tsv file}
}
\value{
Boolean of whether barcodes in tsv file matches colnames in SingleCellExperiment object
}
\description{
Checks if barcodes in SingleCellExperiment object matches the saved barcodes file
}
