% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/barcodes.R
\name{barcodes_exists}
\alias{barcodes_exists}
\title{Check if barcodes file exists}
\usage{
barcodes_exists(path, filename = "barcodes.tsv")
}
\arguments{
\item{path}{Path to multisce folder for the object}

\item{filename}{Name of resulting tsv file}
}
\value{
Boolean of whether barcode file exists
}
\description{
Checks if barcodes file exists
}
