% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/barcodes.R
\name{barcodes_load}
\alias{barcodes_load}
\title{Load barcode data}
\usage{
barcodes_load(path, filename = "barcodes.tsv", missing_ignore = FALSE)
}
\arguments{
\item{path}{Path to multisce folder for the object}

\item{filename}{Name of resulting tsv file}

\item{missing_ignore}{Booling indicating if missing barcode file should be ignored (if so, returns empty vector instead of error)}
}
\value{
Vector of barcodes from tsv file
}
\description{
Load barcodes from tsv file
}
