% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/barcodes.R
\name{barcodes_save}
\alias{barcodes_save}
\title{Save barcode data}
\usage{
barcodes_save(
  sce,
  path = multisce_path(sce),
  filename = "barcodes.tsv",
  overwrite = TRUE
)
}
\arguments{
\item{sce}{SingleCellExperiment object}

\item{path}{Path to multisce folder for the object}

\item{filename}{Name of resulting tsv file}
}
\value{
DataFrame containing only the columns with the indicated prefix
}
\description{
Saves barcodes from SingleCellExperiment as separate file
}
