% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coldata.R
\name{coldata_load}
\alias{coldata_load}
\title{Load column data}
\usage{
coldata_load(
  sce = NULL,
  path,
  filename = "coldata",
  coldata_column_prefix = paste0(mainExpName(sce), "_")
)
}
\arguments{
\item{sce}{SingleCellExperiment object}

\item{path}{Path to multisce folder for the object}

\item{filename}{Name of resulting rds file}

\item{coldata_column_prefix}{Column names with this prefix are omitted when saving but included in the returned DataFrame}
}
\value{
DataFrame containing only the columns with the indicated prefix
}
\description{
Loads colData DataFrame from separate file
}
