% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filtering.R
\name{filter_by_coldata}
\alias{filter_by_coldata}
\title{Filter all multisce objects by coldata}
\usage{
filter_by_coldata(
  path,
  filename = "coldata",
  barcode_match = "^(.*)$",
  debug = FALSE
)
}
\arguments{
\item{path}{Path to multisce folder for the object}

\item{filename}{Name of coldata rds file}

\item{barcode_match}{Regexp expression to use on barcodes before matching}

\item{debug}{Debug mode (will print relevant info during filtering)}
}
\description{
Filters all associated objects in an multisce object so their barcodes corresponds to the barcodes in the coldata
}
