% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metadata.R
\name{metadata_entry_load}
\alias{metadata_entry_load}
\title{Load individual metadata entries}
\usage{
metadata_entry_load(entry_name, path, folder = "metadata")
}
\arguments{
\item{entry_name}{Name of metadata entry to be loaded}

\item{path}{Path to multisce folder for the object}

\item{folder}{Subfolder to use for metadata objects}
}
\value{
Metadata entry (list element)
}
\description{
Load individual metadata entries
}
