% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metadata.R
\name{metadata_entry_save}
\alias{metadata_entry_save}
\title{Save individual metadata entries}
\usage{
metadata_entry_save(entry_name, entry, path, folder = "metadata")
}
\arguments{
\item{entry_name}{Name of metadata entry to be saved}

\item{entry}{metadata entry (list element)}

\item{path}{Path to multisce folder for the object}

\item{folder}{Subfolder to use for reducedDim objects}
}
\description{
Save individual metadata entries
}
