% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metadata.R
\name{metadata_list}
\alias{metadata_list}
\title{List available metadata entries}
\usage{
metadata_list(path, folder = "metadata", extension = ".rds|.tsv")
}
\arguments{
\item{path}{Path to multisce folder for the object}

\item{folder}{Subfolder to use for metadata objects}
}
\description{
List available metadata entries
}
