% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metadata.R
\name{metadata_load}
\alias{metadata_load}
\title{Load metadata}
\usage{
metadata_load(
  path,
  metadata_include = "all",
  metadata_exclude = c("multisce_path")
)
}
\arguments{
\item{path}{Path to multisce folder (uses multisce_path from metadata by default)}

\item{metadata_include}{Metadata entries to load ("all" includes all entries)}

\item{metadata_exclude}{Metadata entries to exclude from being load}
}
\description{
Load metadata
}
