% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metadata.R
\name{metadata_save}
\alias{metadata_save}
\title{Save metadata}
\usage{
metadata_save(
  sce,
  path = multisce_path(sce),
  metadata_include = "all",
  metadata_exclude = c("multisce_path")
)
}
\arguments{
\item{sce}{SingleCellElement}

\item{path}{Path to multisce folder (uses multisce_path from metadata by default)}

\item{metadata_include}{Metadata entries to save ("all" includes all entries)}

\item{metadata_exclude}{Metadata entries to exclude from being saved}
}
\value{
List of metadata entries excluded from being saved
}
\description{
Save metadata
}
