% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multisce_individual.R
\name{multisce_individual_save}
\alias{multisce_individual_save}
\title{Save file}
\usage{
multisce_individual_save(object, path, filename, extension = ".rds")
}
\arguments{
\item{object}{Object to be saved}

\item{path}{Path to multisce folder}

\item{filename}{Filename without extension}

\item{extension}{Filename extension}
}
\description{
For saving each of the separate parts of the multisce in the save format (currently RDS)
}
