% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multisce.R
\name{multisce_path<-}
\alias{multisce_path<-}
\title{Set current multisce path}
\usage{
multisce_path(sce, ...) <- value
}
\arguments{
\item{sce}{SingleCellExperiment object for main experiment.}

\item{value}{Path to multisce folder}
}
\value{
SingleCellExperiment object with multisce_path set
}
\description{
Set current multisce path
}
