% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multisce.R
\name{multisce_path}
\alias{multisce_path}
\title{Get current multisce path}
\usage{
multisce_path(sce)
}
\arguments{
\item{sce}{SingleCellExperiment object for main experiment.}
}
\value{
Path to multisce folder
}
\description{
Get current multisce path
}
