% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multisce.R
\name{multisce_save}
\alias{multisce_save}
\title{Save SCE and conneced altExp into multisce folder}
\usage{
multisce_save(
  sce,
  path,
  main_name = SingleCellExperiment::mainExpName(sce),
  altexp_include = "all",
  altexp_exclude = c(),
  reduceddim_include = "all",
  reduceddim_exclude = c(),
  rownames_strip_prefix = TRUE,
  rownames_prefix_sep = "_",
  main_prefix = paste0(main_name, "__")
)
}
\arguments{
\item{sce}{SingleCellExperiment object}

\item{path}{Path to multisce folder for the object}

\item{main_name}{Name of main experiment (defaults to mainExpName of sce)}

\item{altexp_include}{If "all", all altExps are saved (unless specified in altexp_exclude). Otherwise, a vector of altExp names to be saved}

\item{altexp_exclude}{Vector of altExp names to exclude from being saved}

\item{reduceddim_include}{If "all", all reducedDims are saved (unless specified in reduceddim_exclude). Otherwise, a vector of reducedDim names to be saved}

\item{reduceddim_exclude}{Vector of reducedDim names to exclude from being saved}

\item{rownames_strip_prefix}{Should rowname prefix (altExp name) be stripped before saving?}

\item{rownames_prefix_sep}{Separator used for pasting rownames prefix}

\item{main_prefix}{Names with this prefix are omitted when saving the colData and reducedDims separately but included in the colData of the saved main SCE}
}
\description{
Saves the main SCE and its connected alternative experiments (altExp) into
separate RDS files in the multisce folder
}
