% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reduceddim.R
\name{reduceddim_list}
\alias{reduceddim_list}
\title{List available reducedDims}
\usage{
reduceddim_list(path, folder = "reducedDim", extension = ".rds")
}
\arguments{
\item{path}{Path to multisce folder for the object}

\item{folder}{Subfolder to use for reducedDim objects}
}
\description{
List available reducedDims
}
