% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reduceddim.R
\name{reduceddim_load}
\alias{reduceddim_load}
\title{Load individual reducedDim}
\usage{
reduceddim_load(reducedDimName, path, folder = "reducedDim")
}
\arguments{
\item{reducedDimName}{Name of reducedDim to be loaded}

\item{path}{Path to multisce folder for the object}

\item{folder}{Subfolder to use for reducedDim objects}
}
\description{
Load individual reducedDim
}
