% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reduceddim.R
\name{reduceddim_save}
\alias{reduceddim_save}
\title{Save individual reducedDim}
\usage{
reduceddim_save(reducedDimName, reduceddim, path, folder = "reducedDim")
}
\arguments{
\item{reducedDimName}{Name of reducedDim to be saved}

\item{reduceddim}{reducedDim object}

\item{path}{Path to multisce folder for the object}

\item{folder}{Subfolder to use for reducedDim objects}
}
\description{
Save individual reducedDim
}
