% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sce.R
\name{sce_list}
\alias{sce_list}
\title{List available SingleCellExperiments}
\usage{
sce_list(path, folder = "sce", extension = ".rds")
}
\arguments{
\item{path}{Path to multisce folder for the object}

\item{folder}{Subfolder within path where sce are located}

\item{extension}{File extension used for individual sce files}
}
\description{
List available SingleCellExperiments
}
