% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sce.R
\name{sce_load}
\alias{sce_load}
\title{Load SingleCellExperiment}
\usage{
sce_load(path, filename, folder = "sce")
}
\arguments{
\item{path}{Path to multisce folder for the object}

\item{filename}{Name of sce}

\item{folder}{Subfolder within path where sce are located}
}
\description{
Load SingleCellExperiment
}
