% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sce.R
\name{sce_save}
\alias{sce_save}
\title{Save SingleCellExperiment}
\usage{
sce_save(sce, path = multisce_path(sce), filename, folder = "sce")
}
\arguments{
\item{sce}{SingleCellExperiment object}

\item{path}{Path to multisce folder for the object}

\item{filename}{Name of sce}

\item{folder}{Subfolder within path where sce are located}
}
\description{
Save SingleCellExperiment
}
