% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multisce_individual.R
\name{multisce_individual_load}
\alias{multisce_individual_load}
\title{Load file}
\usage{
multisce_individual_load(path, filename, extension = ".rds")
}
\arguments{
\item{path}{Path to multisce folder}

\item{filename}{Filename without extension}

\item{extension}{Filename extension}
}
\description{
For loading the separate parts of the multisce in the saved format (currently RDS)
}
