% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reduceddim.R
\name{reduceddim_save}
\alias{reduceddim_save}
\title{Save individual reducedDim}
\usage{
reduceddim_save(
  reducedDimName,
  reduceddim,
  path,
  folder = "reducedDim",
  barcodes_file = "barcodes.tsv",
  barcodes_overwrite = FALSE
)
}
\arguments{
\item{reducedDimName}{Name of reducedDim to be saved}

\item{reduceddim}{reducedDim object}

\item{path}{Path to multisce folder for the object}

\item{folder}{Subfolder to use for reducedDim objects}

\item{barcodes_file}{Name of file to save barcodes (default: barcodes.tsv)}

\item{barcodes_overwrite}{Should barcodes file be overwritten if existing - may break colname relationship to objects not currently loaded in the multisce}
}
\description{
Save individual reducedDim
}
